refactor(raw): color, name 원시값 포장
