+++ src/main/java/nextstep/subway/constant/Message.java
@@ -9,5 +9,6 @@ public class Message {
     public final static String NOT_VALID_DUPLICATED_SECTION_STATIONS = "상행역과 하행역이 이미 노선에 모두 등록되어 있어 구간을 등록할 수 없습니다.";
     public final static String NOT_VALID_ANY_STATION = "상행역과 하행역 둘 중 하나라도 포함되어야 구간을 등록할 수 있습니다.";
     public final static String NOT_VALID_UNDER_ZERO_DISTANCE = "거리는 1이상이어야 합니다.";
+    public final static String NOT_VALID_EMPTY = "빈 값은 입력할 수 없습니다.";
 
 }
