+++ src/main/java/nextstep/subway/domain/Station.java
@@ -1,5 +1,7 @@
 package nextstep.subway.domain;
 
+import nextstep.subway.domain.raw.Name;
+
 import javax.persistence.*;
 
 @Entity
@@ -8,14 +10,14 @@ public class Station extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(unique = true)
-    private String name;
+    @Embedded
+    private Name name;
 
     public Station() {
     }
 
     public Station(String name) {
-        this.name = name;
+        this.name = new Name(name);
     }
 
     public Long getId() {
@@ -23,6 +25,6 @@ public class Station extends BaseEntity {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 }
