+++ src/main/java/nextstep/subway/domain/raw/Name.java
@@ -1,0 +1,24 @@
+package nextstep.subway.domain.raw;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+import java.util.Objects;
+
+@Embeddable
+public class Name {
+    @Column(nullable = false)
+    private String name;
+
+    protected Name() {
+    }
+
+    public Name(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+
+}
